requirements-dev.txt
# Development dependencies for Ansible Custom EE Builder

# Core Ansible tools
ansible-core>=2.15.0
ansible-builder>=3.0.0
ansible-navigator>=3.0.0

# Container runtime support
podman-py>=4.6.0

# YAML processing
PyYAML>=6.0

# HTTP requests
requests>=2.31.0

# Development tools
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.2.0

# Security scanning
safety>=2.3.0
bandit>=1.7.0